restructure folders and files, each module in its own file
