+++ src/index.js
@@ -1,155 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-
-function Square(props) {
-  return (
-    <button
-      className="square"
-      onClick={ props.onClick }>
-      {props.value}
-    </button>
-  );
-}
-  
-class Board extends React.Component {
-
-  renderSquare(i) {
-    return <Square
-              value={this.props.squares[i]}
-              onClick={() => this.props.onClick(i)}
-           />;
-  }
-  
-  render() {
-    return (
-      <div>
-        <div className="board-row">
-          {this.renderSquare(0)}
-          {this.renderSquare(1)}
-          {this.renderSquare(2)}
-        </div>
-        <div className="board-row">
-          {this.renderSquare(3)}
-          {this.renderSquare(4)}
-          {this.renderSquare(5)}
-        </div>
-        <div className="board-row">
-          {this.renderSquare(6)}
-          {this.renderSquare(7)}
-          {this.renderSquare(8)}
-        </div>
-      </div>
-    );
-  }
-}
-  
-class Game extends React.Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      history: [{
-        squares: Array(9).fill(null),
-        coordinates: [null, null],
-      }],
-      stepNumber: 0,
-      xIsNext: true,
-    }
-  }
-
-  handleClick(i) {
-    const history = this.state.history.slice(0, this.state.stepNumber + 1);
-    const current = history[history.length - 1];
-    const squares = current.squares.slice();
-
-    if (squares[i] || this.calculateWinnder(squares)) {
-      return;
-    }
-
-    squares[i] = this.state.xIsNext ? 'X' : 'O';
-    this.setState({
-      history: history.concat([{
-        squares: squares,
-        coordinates: this.coordinates(i),
-      }]),
-      stepNumber: history.length,
-      xIsNext: !this.state.xIsNext,
-    });
-  }
-
-  calculateWinnder(squares) {
-    const lines = [
-      [0, 1, 2],
-      [3, 4, 5],
-      [6, 7, 8],
-      [0, 3, 6],
-      [1, 4, 7],
-      [2, 5, 8],
-      [0, 4, 8],
-      [2, 4, 6],
-    ];
-  
-    for (let i = 0; i < lines.length; i++) {
-      const [a, b, c] = lines[i];
-      if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-        return squares[a]
-      }
-    }
-    return null;
-  }
-  
-  coordinates(i) {
-    const x = Math.floor(i / 3);
-    const y = i % 3;
-    return [x, y];
-  }  
-
-  jumpTo(step) {
-    this.setState({
-      stepNumber: step,
-      xIsNext: (step % 2) === 0,
-    });
-  }
-
-  render() {
-    const history = this.state.history;
-    const current = history[this.state.stepNumber];
-    const winner = this.calculateWinnder(current.squares);
-
-    const moves = history.map((step, move) => {
-      let description = move ? 'Go to move #' + move : 'Go to game start';
-      if (move) description += '(' + step.coordinates[0] + ', ' + step.coordinates[1] + ')';
-      return (
-        <li key={move}>
-          <button onClick={() => this.jumpTo(move)}>{description}</button>
-        </li>
-      );
-    });
-
-    let status = '';
-    if (winner) {
-      status = 'Winnser is ' + winner;
-    } else {
-      status = 'Next player: ' + (this.state.xIsNext ? 'X' : 'O');
-    }
-
-    return (
-      <div className="game">
-        <div className="game-board">
-          <Board
-            squares={current.squares}
-            onClick={(i) => this.handleClick(i)}
-          />
-        </div>
-        <div className="game-info">
-          <div className="status">{status}</div>
-          <ol>{moves}</ol>
-        </div>
-      </div>
-    );
-  }
-}
+import Game from './components/Game';
   
 // ========================================
